order progress

+++ Project/DataModels/AccountModel.cs
@@ -1,7 +1,7 @@
 public class AccountModel
 {
 
-    public Int64 Id { get; set; }
+    public int Id { get; set; }
     public string EmailAddress { get; set; }
 
     public string Password { get; set; }
@@ -12,7 +12,7 @@ public class AccountModel
 
     public AccountModel(Int64 id, string email, string password, string? fullname, Int64 admin)
     {
-        Id = id;
+        Id = (int)id;
         EmailAddress = email;
         Password = password;
         FullName = fullname;
